+++ src/legwork/__init__.py
@@ -1,4 +1,5 @@
 from . import evol, psd, snr, source, strain, utils, visualisation
+from ._logging import logger
 from ._version import __version__
 
-__all__ = ["evol", "psd", "snr", "source", "strain", "utils", "visualisation", "__version__"]
+__all__ = ["evol", "psd", "snr", "source", "strain", "utils", "visualisation", "logger", "__version__"]